Code reformatting and clean up
