+++ backend/.gitignore
@@ -1,3 +1,3 @@
 venv
 .env
-__pycache__
+__pycache__/
